refactoring view pager
