+++ app/src/main/java/com/taboola/taboolasample/samples/tabs/BaseTaboolaFragment.java
@@ -1,0 +1,8 @@
+package com.taboola.taboolasample.samples.tabs;
+
+import android.support.v4.app.Fragment;
+
+public abstract class BaseTaboolaFragment extends Fragment {
+
+    public abstract void onPageSelected();
+}
