+++ app/src/main/res/layout/fragment_lv_sample.xml
@@ -3,6 +3,7 @@
     android:id="@+id/feed_lv"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:layout_marginLeft="16dp"
     android:layout_marginBottom="10dp"
-    android:layout_marginRight="16dp" />
+    android:layout_marginLeft="16dp"
+    android:layout_marginRight="16dp"
+    android:layout_marginTop="10dp" />
