+++ app/src/main/res/values/strings.xml
@@ -47,8 +47,18 @@
         Hercle, coordinatae camerarius!.
         Travel heavily like a cloudy lagoon.
         Tremble, scotty, mystery!
-
+        Favere sensim ducunt ad audax boreas.
+        Galataes studere in talis piscinam!
+        When simmering bitter chicken, be sure they are room temperature.
+        Belay, love!
+        Queens are the tribbles of the ancient sonic shower.
+        Pol, a bene sectam, grandis tabes!
+        The fish ransacks with fortune, haul the fortress.
+        Where is the stormy seashell?
+        O, haul me mast, ye golden wind!
+        All of those transformators control bare, dead admirals.
     </string>
+
     <string name="endless_feed">Endless Feed</string>
     <string name="required">Required</string>
     <string name="rv_sample">RecyclerViewSample</string>
